+++ rpiTempMod/RpiTempmonGraph.py
@@ -1,110 +1,135 @@
 #!/usr/bin/env python3
 """
-module imported into rpi_tempmon contains functions related
-containg a class with methods to display graphs from matlib
+Module imported into rpi_tempmon contains
+containing a class with methods to display graphs from matlib
+Class : MatplotGraph
 """
 import os
+import re
+import dateutil
 import matplotlib as mpl
 mpl.use('tkagg')
 import matplotlib.pyplot as plt
 import matplotlib.dates as md
-import dateutil
-from rpiTempMod import RpiTempmonWork as Work  # for data log func
+# My modules
+from . import RpiTempmonWork as Work
+
 
 
 class MatplotGraph(object):
-    """class with 2 methods to display graphs with matplotlib"""
+    """
+    class with 6 methods to display graphs with matplotlib
+    Methods:
+    (1) init: pass
+    (2) graph_log_data: Define data to draw graphs based on user choice
+    (3) display_menu: Method to display a menu for
+        user to select graph type
+    (4) draw_graph: Draw a  graph from logdata
+    (5) graph_live_data: Draw live data graphs
+    (6) Plot_now: Called from method graph_live_data to draw graph
+    """
+
     def __init__(self, name):
-        """docstring"""
+        """init class with name and define self.selection variable"""
         self.name = name
+        self.selection = 0
 
     def graph_log_data(self, destlog):
-        """draw a  graph of pi GPU CPU from logdata"""
+        """define data to draw graphs from logdata"""
         # display menu return user selection
-        selection = self.display_menu()
+        self.display_menu()
 
         # get data from log file function data_func
-        timelist, cpulist, gpulist, cpu_uselist, ramlist, swaplist\
+        timelist, unixlist, cpulist, gpulist, cpu_uselist, ramlist, swaplist\
             = Work.data_func(destlog, True)
+        if re.match('[1-4]', self.selection):
+            yaxislist = timelist
+        elif re.match('[5-8]', self.selection):
+            yaxislist = unixlist
 
-        # based on selection define y axis , title and call graph function
-        # passing in data
-        if selection == '1':
+        if (self.selection == '1' or self.selection == '5'):
             plotlabel1 = 'CPU'
             plotlabel2 = 'GPU'
             ylabel = 'Temperature (degrees)'
             title = 'CPU & GPU temperature of RPi'
-            self.draw_graph(timelist, cpulist, gpulist,
+            self.draw_graph(yaxislist, cpulist, gpulist,
                             plotlabel1, plotlabel2, ylabel, title)
-        elif selection == '2':
+
+        elif (self.selection == '2' or self.selection == '6'):
             plotlabel1 = 'CPU temp'
             plotlabel2 = 'CPU usage'
             ylabel = 'CPU usage(%) / Temperature (degrees)'
             title = 'CPU temperature and usage RPi'
-            self.draw_graph(timelist, cpulist, cpu_uselist,
+            self.draw_graph(yaxislist, cpulist, cpu_uselist,
                             plotlabel1, plotlabel2, ylabel, title)
-        elif selection == '3':
+        elif (self.selection == '3' or self.selection == '7'):
             plotlabel1 = 'RAM'
             plotlabel2 = 'SWAP'
             ylabel = 'Memory (% used)'
             title = 'RAM & Swap memory usage of RPi'
-            self.draw_graph(timelist, ramlist, swaplist,
+            self.draw_graph(yaxislist, ramlist, swaplist,
                             plotlabel1, plotlabel2, ylabel, title)
-        elif selection == '4':
+        elif (self.selection == '4' or self.selection == '8'):
             plotlabel1 = 'CPU'
             ylabel = 'Memory (% used)'
             title = 'CPU usage of RPi'
-            self.draw_graph(timelist, cpu_uselist, False,
+            self.draw_graph(yaxislist, cpu_uselist, False,
                             plotlabel1, False, ylabel, title)
         else:
-            print("Error Bad selection value")
-            quit()
+            Work.msg_func("red", "Error: graph_log_data: Bad selection value")
+            return
 
     def display_menu(self):
         """ method to display a menu for
         user to select graph"""
+        os.system('clear')
         menu = []
         menu.append("CPU and GPU Temperature versus Time-date")
         menu.append("CPU Temperature and CPU usage versus Time-date")
         menu.append("RAM and Swap memory usage versus Time-date")
         menu.append("CPU usage versus Time-date")
+        menu.append("CPU and GPU Temperature versus Epoch time")
+        menu.append("CPU Temperature and CPU usage versus Epoch time")
+        menu.append("RAM and Swap memory usage versus Epoch time")
+        menu.append("CPU usage versus Epoch time")
         menu.append("CPU usage versus live Time")
         menu.append("GPU Temperature versus live Time")
+        menu.append("RAM usage versus live Time")
         menu.append("Exit")
-        while True:
-            print("\n")
-            print(23 * "-", "GRAPH MENU OPTIONS", 23 * "-")
-            for number, string in enumerate(menu):
-                print(number+1, string)
-            print(67 * "-")
-            print("\n")
-            selection = input("Please Select:")
-            if selection == '1':
-                return selection
-            elif selection == '2':
-                return selection
-            elif selection == '3':
-                return selection
-            elif selection == '4':
-                return selection
-            elif selection == '5':
-                self.graph_live_data("CPU")
-                break
-            elif selection == '6':
-                self.graph_live_data("GPU")
-                break
-            elif selection == '7':
-                break
-            else:
-                print("\n\t ** Warning : Unknown Option Selected! **")
-        quit()
+        try:
+            while True:
+                print("\n")
+                Work.msg_func("blue", "Graph Menu Options")
+                Work.msg_func("line", "")
+                for number, string in enumerate(menu):
+                    print(number+1, string)
+                Work.msg_func("line", "")
+                self.selection = (input("Please Select:"))
+                if int(self.selection) <= 8:
+                    return
+                elif self.selection == '9':
+                    self.graph_live_data("CPU")
+                    break
+                elif self.selection == '10':
+                    self.graph_live_data("GPU")
+                    break
+                elif self.selection == '11':
+                    self.graph_live_data("RAM")
+                    break
+                elif self.selection == '12':
+                    quit()
+                else:
+                    Work.msg_func("red", "\n ** Warning : Unknown Option Selected! **")
+                    Work.msg_func("anykey", "")
+                    os.system('clear')
+        except ValueError as error:
+            print(error)
+            Work.msg_func("red", "Error: Wrong menu Input: Integer only : Try Again")
+            quit()
 
     def draw_graph(self, timelist, yaxis_list1,
                    yaxis_list2, plot_label1, plot_label2, yaxis_label, graph_title):
         """ Method to draw graphs two  modes, single and doulbe yaxis """
-
-        # parse dates format
-        mydates = [dateutil.parser.parse(s) for s in timelist]
         # convert to ints as strings cause issue with graph in new matlib version
         yaxis_list1 = list(map(float, yaxis_list1))
 
@@ -112,12 +137,23 @@ class MatplotGraph(object):
             yaxis_list2 = list(map(float, yaxis_list2))
 
         plt.xticks(rotation=90)
-        plt.xticks(fontsize=5)
+        plt.xticks(fontsize=6)
         plt.subplots_adjust(bottom=0.2)
         axisx = plt.gca()
-        axisx.set_xticks(mydates)
-        xfmt = md.DateFormatter('%m/%d %H:%M')
-        axisx.xaxis.set_major_formatter(xfmt)
+        # check user input for time date or unix epoch for yaxis
+        if self.selection == 0:
+            mydates = timelist
+            plt.xlabel('TestRuns')
+        elif re.match('[1-4]', self.selection):
+            mydates = [dateutil.parser.parse(s) for s in timelist]
+            axisx.set_xticks(mydates)
+            xfmt = md.DateFormatter('%m/%d %H:%M')
+            axisx.xaxis.set_major_formatter(xfmt)
+            plt.xlabel('Date time stamp (DD-MM HH:MM)')
+        elif re.match('[5-8]', self.selection):
+            mydates = timelist
+            plt.xlabel('Unix epoch time (seconds)')
+
         axisx.xaxis.label.set_color('red')
         axisx.yaxis.label.set_color('red')
 
@@ -126,7 +162,7 @@ class MatplotGraph(object):
         if plot_label2:  # single plot graph mode
             plt.plot(mydates,
                      yaxis_list2, label=plot_label2, marker='*')
-        plt.xlabel('Date time stamp (DD-MM HH:MM)')
+
         plt.ylabel(yaxis_label)
         plt.title(graph_title, color='green')
         plt.legend(loc='upper right', fancybox=True, shadow=True)
@@ -145,21 +181,28 @@ class MatplotGraph(object):
                 time_axis.append(.5)
 
             while True:
-                # GPU
+
                 time_axis.append(yaxis_data)
                 time_axis.pop(0)
                 if choice == "GPU":
                     ostemp = os.popen('vcgencmd measure_temp').readline()
                     yaxis_data = (ostemp.replace("temp=", "").replace("'C\n", ""))
-                    labels = ("RPi GPU temp", "Temperature ('C)", "GPU")
+                    labels = (" GPU live temp", "Temperature ('C)", "GPU")
                     yaxis_data = float(yaxis_data)
                     time_axis.append(yaxis_data)
                     time_axis.pop(0)
                     self.plot_now(time_axis, labels)
-                else:
+                elif choice == 'CPU':
                     yaxis_data = Work.get_cpu_use()
                     yaxis_data = float(yaxis_data)
-                    labels = ("RPi CPU usage", "Usage (%)", "CPU")
+                    labels = (" CPU live usage", "Usage (%)", "CPU")
+                    time_axis.append(yaxis_data)
+                    time_axis.pop(0)
+                    self.plot_now(time_axis, labels)
+                else:  # RAM
+                    yaxis_data = Work.get_ram_info()
+                    yaxis_data = float(yaxis_data)
+                    labels = (" RAM live usage", "Usage (%)", "RAM")
                     time_axis.append(yaxis_data)
                     time_axis.pop(0)
                     self.plot_now(time_axis, labels)
@@ -167,7 +210,8 @@ class MatplotGraph(object):
                 plt.pause(2)
         except Exception as error:
             print(error)
-            print("Real-time matplotlib plot shutdown")
+            Work.msg_func("bold", "Real-time matplotlib plot shutdown")
+            quit()
 
     def plot_now(self, timeaxis, labels):
         """ Called from method graph_live_data to draw graph"""
@@ -178,7 +222,7 @@ class MatplotGraph(object):
         plt.ylim([1, 100])
         plt.ylabel(y_label, color='red')
 
-        plt.title(title, color='green')
+        plt.title(self.name + title, color='green')
         plt.grid(True)
         plt.xlabel("Time (last 300 seconds)", color='red')
         plt.plot(timeaxis, color='blue', marker='*', label=plot_label)
@@ -188,8 +232,8 @@ class MatplotGraph(object):
 
 def importtest(text):
     """import print test statement"""
-    pass
     # print(text)
+    pass
 
 # ===================== MAIN ===============================